2023/02/23

+++ ft_functions/ft_math/ft_pow.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-long long
+double
 	ft_pow(register double base, register double power)
 {
 	register int	i;
